backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.1.13
pydantic==2.5.0
PyYAML==6.0.1
python-multipart==0.0.6